// payload-types.ts
export interface LandingPage {
  id: string;
  title: string;
  slug: string;
  components?:
    | (
        | {
            heroImage: string | Media;
            heroTitle: string;
            heroSubtitle?: string | null;
            heroBody?: string | null;
            id?: string | null;
            blockName?: string | null;
            blockType: 'hero';
          }
        | {
            type: 'single' | 'multiple';
            title?: string | null;
            body?: string | null;
            items: {
              title?: string | null;
              body?: string | null;
              id?: string | null;
              blockName?: string | null;
              blockType: 'accordion-item';
            }[];
            id?: string | null;
            blockName?: string | null;
            blockType: 'accordion';
          }
      )[]
    | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: string;
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
}